add cmd
