+++ src/lib.rs
@@ -1,4 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+pub mod config;
+pub mod server;
+
 pub use {mcp_server, wassette};
